+++ eq-author/src/components/AnswerContent/AnswerProperties/DateProperties.js
@@ -1,17 +1,13 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { enableOn } from "utils/featureFlags";
-
 import { Select } from "components/Forms";
 import Required from "components/AnswerContent/Required";
-
 import MultiLineField from "components/AnswerContent/Format/MultiLineField";
-
 import styled from "styled-components";
 
-const Container = styled.div`
-  display: flex;
-  align-items: center;
+const StyledSelect = styled(Select)`
+  width: 12em;
 `;
 
 const monthText = enableOn(["hub"]) ? "mm" : "Month";
@@ -27,31 +23,30 @@ const DateProperties = ({ answer, updateAnswer }) => {
       },
     });
   };
+
   return (
     <>
-      <Container>
-        <MultiLineField id="date-format" label={"Date type"}>
-          <Select
-            data-test="select"
-            value={answer.properties.format}
-            onChange={onUpdateFormat}
-            id="date-format"
-            name="format"
-          >
-            <option
-              data-test="day-month-year"
-              value="dd/mm/yyyy"
-            >{`dd / ${monthText} / yyyy`}</option>
-            <option
-              data-test="month-year"
-              value="mm/yyyy"
-            >{`${monthText} / yyyy`}</option>
-            <option data-test="year" value="yyyy">
-              yyyy
-            </option>
-          </Select>
-        </MultiLineField>
-      </Container>
+      <MultiLineField id="date-format" label={"Date type"}>
+        <StyledSelect
+          data-test="select"
+          value={answer.properties.format}
+          onChange={onUpdateFormat}
+          id="date-format"
+          name="format"
+        >
+          <option
+            data-test="day-month-year"
+            value="dd/mm/yyyy"
+          >{`dd / ${monthText} / yyyy`}</option>
+          <option
+            data-test="month-year"
+            value="mm/yyyy"
+          >{`${monthText} / yyyy`}</option>
+          <option data-test="year" value="yyyy">
+            yyyy
+          </option>
+        </StyledSelect>
+      </MultiLineField>
       <Required answer={answer} updateAnswer={updateAnswer} />
     </>
   );
