+++ eq-author/src/components/AnswerContent/AnswerProperties/RadioProperties.js
@@ -1,21 +1,9 @@
 import React from "react";
 import PropTypes from "prop-types";
 import Required from "components/AnswerContent/Required";
-import styled from "styled-components";
-
-const Container = styled.div`
-  display: flex;
-  align-items: center;
-`;
 
 const RadioProperties = ({ answer, updateAnswer }) => {
-  return (
-    <>
-      <Container>
-        <Required answer={answer} updateAnswer={updateAnswer} />
-      </Container>
-    </>
-  );
+  return <Required answer={answer} updateAnswer={updateAnswer} />;
 };
 
 RadioProperties.propTypes = {
