+++ eq-author/src/components/AnswerContent/AnswerProperties/UnitProperties.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import MultiLineField from "components/AnswerContent/AnswerProperties/Format/MultiLineField";
+import MultiLineField from "components/AnswerContent/Format/MultiLineField";
 import Decimal from "components/AnswerContent/Decimal";
 import Required from "components/AnswerContent/Required";
 import { Autocomplete } from "components/Autocomplete";
@@ -12,8 +12,7 @@ import { unitPropertyErrors } from "constants/validationMessages";
 import ValidationError from "components/ValidationError";
 
 const Container = styled.div`
-  display: flex;
-  align-items: center;
+  width: 15em;
 `;
 
 const filterCondition = (x, query) =>
@@ -93,21 +92,17 @@ const UnitProperties = ({ answer, page, updateAnswer, updateAnswerOfType }) => {
           {unitPropertyErrors[errors[0].errorCode].message}
         </ValidationError>
       )}
-      <Container>
-        <MultiLineField id="decimal" htmlFor="decimal" label="Decimal places">
-          <Decimal
-            id={answer.id}
-            answer={answer}
-            page={page}
-            data-test="decimals"
-            updateAnswerOfType={updateAnswerOfType}
-            value={answer.properties.decimals}
-          />
-        </MultiLineField>
-      </Container>
-      <Container>
-        <Required answer={answer} updateAnswer={updateAnswer} />
-      </Container>
+      <MultiLineField id="decimal" htmlFor="decimal" label="Decimal places">
+        <Decimal
+          id={answer.id}
+          answer={answer}
+          page={page}
+          data-test="decimals"
+          updateAnswerOfType={updateAnswerOfType}
+          value={answer.properties.decimals}
+        />
+      </MultiLineField>
+      <Required answer={answer} updateAnswer={updateAnswer} />
     </>
   );
 };
