+++ eq-author/src/components/AnswerContent/FallbackProperty/Fallback.js
@@ -2,8 +2,8 @@ import React from "react";
 import styled from "styled-components";
 import PropTypes from "prop-types";
 import CustomPropTypes from "custom-prop-types";
-import InlineField from "components/AnswerContent/AnswerProperties/Format/InlineField";
-import MultiLineField from "components/AnswerContent/AnswerProperties/Format/MultiLineField";
+import InlineField from "components/AnswerContent/Format/InlineField";
+import MultiLineField from "components/AnswerContent/Format/MultiLineField";
 import { SimpleSelect } from "components/Forms/Select";
 import Collapsible from "components/Collapsible";
 import ValidationError from "components/ValidationError";
